test/requirements.txt
pytest
redis
dropbox
python-dotenv
names